feat: game card 반응형 수정 및 파일 정리
